requirements.txt
agate==1.9.1
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
babel==2.17.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
daff==1.4.2
dbt==1.0.0.40.6
dbt-adapters==1.16.7
dbt-common==1.30.0
dbt-core==1.10.11
dbt-extractor==0.6.0
dbt-postgres==1.9.1
dbt-protos==1.0.375
dbt-semantic-interfaces==0.9.0
deepdiff==8.6.1
deprecation==2.1.0
et_xmlfile==2.0.0
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
isodate==0.6.1
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
leather==0.4.0
MarkupSafe==3.0.2
mashumaro==3.14
more-itertools==10.8.0
msgpack==1.1.1
networkx==3.5
numpy==2.3.3
openpyxl==3.1.5
orderly-set==5.5.0
packaging==25.0
pandas==2.3.2
parsedatetime==2.6
pathspec==0.12.1
postgrest==1.1.1
protobuf==6.32.1
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.2
realtime==2.7.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
snowplow-tracker==1.1.0
SQLAlchemy==2.0.43
sqlparse==0.5.3
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
text-unidecode==1.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
zipp==3.23.0
